saving preliminary af model progress
